Refactor monolithic optics.py into a modular package structure
