--- apts/optics.py
@@ -1,1363 +1,0 @@
-import functools
-import math
-import operator
-from typing import TYPE_CHECKING, Any, Optional, Sequence, Union, cast
-
-import numpy
-
-if TYPE_CHECKING:
-    from pint import Quantity
-
-    from .opticalequipment.abstract import OpticalEquipment, OutputOpticalEquipment
-    from .opticalequipment.smart_telescope import SmartTelescope
-
-from .constants import astronomy
-from .opticalequipment.binoculars import Binoculars
-from .opticalequipment.naked_eye import NakedEye
-from .units import get_unit_registry
-from .utils import optics as optics_utils
-
-
-class OpticsUtils:
-    @staticmethod
-    def expand(
-        path: Sequence["OpticalEquipment"],
-    ) -> tuple[
-        Union["OpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-        Sequence[Any],
-        Sequence[Any],
-        Sequence[Any],
-        Sequence[Any],
-        Union["OutputOpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-    ]:
-        from .opticalequipment.barlow import Barlow
-        from .opticalequipment.diagonal import Diagonal
-        from .opticalequipment.filter import Filter
-        from .opticalequipment.reducer.base import Corrector, Flattener, Reducer
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        # First item in the path should be the telescope or binoculars
-        main_optic = path[0]
-        if isinstance(main_optic, (Binoculars, NakedEye, SmartTelescope)):
-            # Treat binoculars as the 'output' as well for path structure consistency
-            return (main_optic, [], [], [], [], main_optic)
-
-        # Original logic for telescopes
-        telescope = main_optic
-        # Last item in the path is output
-        output = cast("OutputOpticalEquipment", path[-1])
-        # Intermediate elements - categorized in a single pass
-        barlows = []
-        diagonals = []
-        filters = []
-        others = []
-
-        for item in path[1:-1]:
-            if isinstance(item, (Barlow, Reducer, Flattener, Corrector)):
-                barlows.append(item)
-            elif isinstance(item, Diagonal):
-                diagonals.append(item)
-            elif isinstance(item, Filter):
-                filters.append(item)
-            else:
-                others.append(item)
-
-        return (telescope, barlows, diagonals, filters, others, output)
-
-    @staticmethod
-    def barlows_multiplications(barlows_list: Sequence[Any]) -> float:
-        barlows = [barlow.magnification for barlow in barlows_list]
-        # Multiply all barlows
-        return float(functools.reduce(operator.mul, barlows, 1))
-
-    @staticmethod
-    def compute_zoom(
-        telescope: Union["OpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-        barlows: Sequence[Any],
-        output: Union["OutputOpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-    ) -> "Quantity":
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return telescope.magnification * get_unit_registry().dimensionless
-
-        # Original logic
-        magnification = OpticsUtils.barlows_multiplications(barlows)
-        return telescope.focal_length * magnification / output._zoom_divider()
-
-    @staticmethod
-    def calculate_airmass(altitude_degrees: float) -> float:
-        """
-        Calculates the relative airmass using the Kasten-Young (1989) formula.
-        Formula: X = 1 / (sin(h) + 0.50572 * (h + 6.07995)^-1.6364)
-        Where h is the apparent altitude in degrees.
-        Source: https://en.wikipedia.org/wiki/Air_mass_(astronomy)
-        """
-        return float(optics_utils.calculate_airmass(altitude_degrees))
-
-    @staticmethod
-    def compute_field_of_view(
-        telescope: Union["OpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-        barlows: Sequence[Any],
-        output: Union["OutputOpticalEquipment", Binoculars, NakedEye, "SmartTelescope"],
-    ) -> "Quantity":
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return telescope.fov()
-
-        # Original logic
-        magnification = OpticsUtils.barlows_multiplications(barlows)
-        zoom = OpticsUtils.compute_zoom(telescope, barlows, output)
-        # TODO: this is not best way to do it
-        return output.field_of_view(telescope, zoom, magnification)
-
-
-class OpticalPath:
-    """
-    Class representing an optical path in a telescope setup.
-    """
-
-    def __init__(self, telescope, barlows, diagonals, filters, others, output=None):
-        self.telescope = telescope
-        self.barlows = barlows
-        self.diagonals = diagonals
-        self.filters = filters
-        if output is None:
-            # Handle 5-argument constructor calls for backward compatibility
-            self.others = []
-            self.output = others
-        else:
-            self.others = others
-            self.output = output
-        # Cache for expensive calculations
-        self._cache = {}
-
-    @classmethod
-    def from_path(cls, path):
-        telescope, barlows, diagonals, filters, others, output = OpticsUtils.expand(
-            path
-        )
-        return cls(telescope, barlows, diagonals, filters, others, output)
-
-    def zoom(self) -> "Quantity":
-        if "zoom" not in self._cache:
-            self._cache["zoom"] = OpticsUtils.compute_zoom(
-                self.telescope, self.barlows, self.output
-            )
-        return self._cache["zoom"]
-
-    def effective_barlow(self) -> float:
-        if "effective_barlow" not in self._cache:
-            self._cache["effective_barlow"] = float(
-                OpticsUtils.barlows_multiplications(self.barlows)
-            )
-        return self._cache["effective_barlow"]
-
-    def label(self) -> str:
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(self.telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return str(self.telescope)
-        return ", ".join(
-            [str(self.telescope)]
-            + [str(item) for item in self.barlows]
-            + [str(item) for item in self.diagonals]
-            + [str(item) for item in self.filters]
-            + [str(item) for item in self.others]
-            + [str(self.output)]
-        )
-
-    def length(self) -> int:
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        # For binoculars, path is [Binoculars], expanded to (Binoculars, [], Binoculars)
-        # So self.barlows is []. Length should be 1 for a direct binocular path.
-        # Original: return 2 + len(self.barlows) (Telescope + Output + Barlows)
-        if isinstance(self.telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return 1  # Just the binoculars itself
-        return (
-            2
-            + len(self.barlows)
-            + len(self.diagonals)
-            + len(self.filters)
-            + len(self.others)
-        )
-
-    def fov(self) -> "Quantity":
-        return OpticsUtils.compute_field_of_view(
-            self.telescope, self.barlows, self.output
-        )
-
-    def fov_width(self) -> "Quantity":
-        return self.output.field_of_view_width(
-            self.telescope, self.zoom(), self.effective_barlow()
-        )
-
-    def fov_height(self) -> "Quantity":
-        return self.output.field_of_view_height(
-            self.telescope, self.zoom(), self.effective_barlow()
-        )
-
-    def fov_diagonal(self) -> "Quantity":
-        return self.output.field_of_view_diagonal(
-            self.telescope, self.zoom(), self.effective_barlow()
-        )
-
-    def is_magnification_useful(self) -> bool:
-        """
-        Checks if the current magnification is within the theoretical useful range
-        of the telescope. For non-visual outputs (cameras), it returns True.
-        """
-        from .opticalequipment.telescope import Telescope
-
-        if (
-            not isinstance(self.telescope, Telescope)
-            or not self.output.is_visual_output()
-        ):
-            return True
-
-        zoom = self.zoom().magnitude
-        return (
-            self.telescope.lowest_useful_magnification()
-            <= zoom
-            <= self.telescope.highest_useful_magnification()
-        )
-
-    def airmass(self, altitude_degrees: float) -> float:
-        """
-        Calculates the relative airmass for a given altitude in degrees.
-        """
-        return float(OpticsUtils.calculate_airmass(altitude_degrees))
-
-    def atmospheric_extinction(
-        self, magnitude: float, altitude_degrees: float, extinction_k: float = 0.2
-    ) -> float:
-        """
-        Calculates the apparent magnitude of an object accounting for atmospheric extinction.
-        Formula: m_apparent = m_zero + k * X
-        Where k is the extinction coefficient and X is the airmass.
-        k typically ranges from 0.15 (very clear) to 0.5+ (hazy/polluted). Default 0.2.
-        """
-        airmass_val = self.airmass(altitude_degrees)
-        return float(
-            optics_utils.calculate_atmospheric_extinction(
-                magnitude, airmass_val, extinction_k
-            )
-        )
-
-    def atmospheric_dispersion(
-        self, altitude_degrees: float, lambda1_nm: float = 400, lambda2_nm: float = 700
-    ) -> "Quantity":
-        """
-        Calculates the atmospheric dispersion (angular separation) between two wavelengths
-        at a given altitude using the Peck and Reeder (1972) refractive index formula.
-        Formula: (n-1) * 10^6 = 64.328 + 29498.1 / (146 - lambda^-2) + 255.4 / (41 - lambda^-2)
-        Dispersion Delta R = (n1 - n2) * tan(zenith_distance)
-        Source: Peck & Reeder (1972), "Refractive Index of Air in the Near Infrared"
-        """
-        dispersion_arcsec = optics_utils.calculate_atmospheric_dispersion(
-            altitude_degrees, lambda1_nm, lambda2_nm
-        )
-        return float(dispersion_arcsec) * get_unit_registry().arcsecond
-
-    def atmospheric_dispersion_in_pixels(
-        self, altitude_degrees: float, lambda1_nm: float = 400, lambda2_nm: float = 700
-    ) -> Optional[float]:
-        """
-        Calculates the atmospheric dispersion in pixels for the current optical path.
-        Useful for planetary imaging to determine if an Atmospheric Dispersion Corrector (ADC) is needed.
-        """
-        dispersion = self.atmospheric_dispersion(
-            altitude_degrees, lambda1_nm, lambda2_nm
-        )
-        scale = self.pixel_scale()
-        if scale is None or scale.magnitude == 0:
-            return None
-        return float(
-            dispersion.to("arcsecond").magnitude / scale.to("arcsecond").magnitude
-        )
-
-    def brightness(self) -> "Quantity":
-        if "brightness" in self._cache:
-            return self._cache["brightness"]
-
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(self.telescope, (Binoculars, NakedEye, SmartTelescope)):
-            # self.telescope.brightness() returns a float like 75.0 (for 75%)
-            # OutputOpticalEquipment.brightness returns a value like <Quantity(75.0, 'dimensionless')>
-            # To be consistent so that .magnitude can be called later:
-            brightness = self.telescope.brightness() * get_unit_registry().dimensionless
-        else:
-            # This already returns a pint Quantity from OutputOpticalEquipment's method
-            brightness = self.output.brightness(self.telescope, self.zoom())
-
-        # Account for filters transmission
-        for f in self.filters:
-            brightness *= f.transmission
-
-        self._cache["brightness"] = brightness
-        return brightness
-
-    def exit_pupil(self) -> "Quantity":
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(self.telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return self.telescope.exit_pupil()  # This returns a Quantity
-
-        # Original logic for telescopes:
-        # telescope.aperture should be a Quantity (e.g., mm)
-        # self.zoom() returns a dimensionless Quantity
-        # The result will be in units of aperture (e.g., mm)
-        if hasattr(self.telescope, "aperture"):
-            zoom_value = self.zoom()
-            if zoom_value.magnitude == 0:  # pyright: ignore
-                return 0 * get_unit_registry().mm
-            return self.telescope.aperture / zoom_value
-
-        # If it's not Binoculars and telescope has no aperture (should not happen for Telescopes)
-        # return a zero quantity to avoid crashes, though this indicates a data problem.
-        return 0 * get_unit_registry().mm
-
-    def elements(self) -> frozenset[Any]:
-        """
-        Return immutable set of elements - used for removing redundant optical paths
-        """
-        elements: set[Any] = set((self.telescope, self.output))
-        elements |= set(self.barlows)
-        elements |= set(self.diagonals)
-        elements |= set(self.filters)
-        elements |= set(self.others)
-        return frozenset(elements)
-
-    def component_list(self) -> list[Any]:
-        """
-        Return ordered list of all optical components in the path.
-        """
-        from .opticalequipment.binoculars import Binoculars
-        from .opticalequipment.naked_eye import NakedEye
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(self.telescope, (Binoculars, NakedEye, SmartTelescope)):
-            return [self.telescope]
-
-        return (
-            [self.telescope]
-            + self.barlows
-            + self.diagonals
-            + self.filters
-            + self.others
-            + [self.output]
-        )
-
-    def total_mass(self) -> "Quantity":
-        if "total_mass" in self._cache:
-            return self._cache["total_mass"]
-
-        from .opticalequipment.abstract import OpticalEquipment
-
-        all_equipment: set[OpticalEquipment] = set()
-        for item in (
-            [self.telescope]
-            + self.barlows
-            + self.diagonals
-            + self.filters
-            + self.others
-            + [self.output]
-        ):
-            if hasattr(item, "collect_all_attached"):
-                item.collect_all_attached(all_equipment)
-            else:
-                all_equipment.add(item)
-
-        total = 0 * get_unit_registry().gram
-        for eq in all_equipment:
-            mass = getattr(eq, "mass", 0 * get_unit_registry().gram)
-            if mass is not None:
-                total += mass
-        self._cache["total_mass"] = total
-        return total
-
-    def backfocus_gap(self) -> Optional["Quantity"]:
-        """
-        Calculate the backfocus gap.
-        Returns a Quantity (distance) or None if no backfocus requirement is defined.
-        """
-        # 1. Find the component that defines the required backfocus
-        required_bf = None
-        start_index = -1
-
-        # Flatten path for searching
-        path = (
-            [self.telescope]
-            + self.barlows
-            + self.diagonals
-            + self.filters
-            + self.others
-            + [self.output]
-        )
-
-        for i, item in enumerate(path):
-            # Check if component has required_backfocus (Reducer, Flattener, etc.)
-            if (
-                hasattr(item, "required_backfocus")
-                and item.required_backfocus is not None
-            ):
-                required_bf = item.required_backfocus
-                start_index = i
-            # For telescopes, 'backfocus' property can also mean required backfocus from flange
-            elif i == 0 and hasattr(item, "backfocus") and item.backfocus is not None:
-                required_bf = item.backfocus
-                start_index = i
-
-        if required_bf is None:
-            return None
-
-        # 2. Calculate actual distance from that component's output to the sensor
-        actual_distance = 0 * get_unit_registry().mm
-        # Sum optical lengths of everything between start and end
-        for item in path[start_index + 1 : -1]:
-            item_ol = getattr(item, "optical_length", 0 * get_unit_registry().mm)
-            if item_ol is not None:
-                actual_distance += item_ol
-
-        # 3. Add the output component's backfocus contribution (e.g. sensor depth)
-        if hasattr(self.output, "backfocus") and self.output.backfocus is not None:
-            actual_distance += self.output.backfocus
-
-        return required_bf - actual_distance
-
-    def get_image_orientation(self):
-        from .opticalequipment.telescope import Telescope
-
-        if not isinstance(self.telescope, Telescope):
-            return (False, False)
-
-        # Start with the telescope's inversion (1 horizontal + 1 vertical flip)
-        flipped_horizontally = True
-        flipped_vertically = True
-
-        # Diagonals are typically used with Refractors and Catadioptrics,
-        # and their effect depends on the telescope type.
-        for diagonal in self.diagonals:
-            if diagonal.is_erecting:
-                # Erecting diagonal adds 1 horizontal and 1 vertical flip
-                flipped_horizontally = not flipped_horizontally
-                flipped_vertically = not flipped_vertically
-            else:
-                # Standard star diagonal adds 1 vertical flip
-                flipped_vertically = not flipped_vertically
-
-        return (flipped_horizontally, flipped_vertically)
-
-    def pixel_scale(self) -> Optional["Quantity"]:
-        if "pixel_scale" in self._cache:
-            return self._cache["pixel_scale"]
-
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            self._cache["pixel_scale"] = None
-            return None
-        # Effective focal length
-        eff_focal_length = self.telescope.focal_length * self.effective_barlow()
-        # Pixel size
-        p_size = self.output.pixel_size()
-        # Formula: (p_size / eff_focal_length) * RAD_TO_ARCSEC
-        scale = (
-            p_size.to("mm").magnitude / eff_focal_length.to("mm").magnitude
-        ) * astronomy.RAD_TO_ARCSEC
-        res = scale * get_unit_registry().arcsecond
-        self._cache["pixel_scale"] = res
-        return res
-
-    def sampling(self, seeing: float) -> Optional[str]:
-        """
-        Calculates the sampling status based on the resolution limit and the pixel scale.
-        The resolution limit is the larger of the atmospheric seeing and the telescope's diffraction limit (Rayleigh limit).
-
-        According to the Nyquist-Shannon sampling theorem, the ideal sampling (critical sampling)
-        is at least 2 pixels per resolution element. In astrophotography, a ratio of 2.0 to 3.0
-        is often considered ideal to capture all available detail without excessive oversampling.
-
-        Thresholds:
-        - Under-sampled: ratio < 1.0 (Information lost, stars look like squares)
-        - Well-sampled: 1.0 <= ratio <= 3.0 (Good balance for most conditions)
-        - Over-sampled: ratio > 3.0 (No extra detail gained, smaller field of view, lower SNR)
-
-        Source: Nyquist-Shannon sampling theorem
-        """
-        scale = self.pixel_scale()
-        if scale is None or scale.magnitude == 0:
-            return None
-
-        # Effective resolution limit is the larger of seeing and diffraction limit
-        r_limit = seeing
-        diffraction_limit = self.rayleigh_limit()
-        if diffraction_limit is not None:
-            r_limit = max(seeing, diffraction_limit.to("arcsecond").magnitude)
-
-        ratio = r_limit / scale.magnitude
-        if ratio < 1.0:
-            return "Under-sampled"
-        elif ratio <= 3.0:
-            return "Well-sampled"
-        else:
-            return "Over-sampled"
-
-    def sampling_status(self, seeing: float = 2.0) -> Optional[str]:
-        """
-        Returns the sampling status as a string for a given seeing in arcseconds.
-        """
-        return self.sampling(seeing)
-
-    def critical_focus_zone(self, wavelength: float = 550) -> Optional["Quantity"]:
-        if not hasattr(self.telescope, "focal_ratio"):
-            return None
-        # wavelength in nm
-        fr = (self.telescope.focal_ratio() * self.effective_barlow()).magnitude
-        # CFZ = 2.44 * (wavelength/1000) * fr^2
-        cfz = 2.44 * (wavelength / 1000.0) * (fr**2)
-        return cfz * get_unit_registry().micrometer
-
-    def thermal_drift(self, delta_t: float) -> Optional["Quantity"]:
-        if (
-            not hasattr(self.telescope, "tube_material")
-            or self.telescope.tube_material is None
-        ):
-            return None
-        # delta_t in Celsius
-        length = self.telescope.focal_length.to("mm").magnitude
-        alpha = self.telescope.tube_material.value  # m/(m*K)
-        drift = length * alpha * delta_t
-        return drift * get_unit_registry().mm
-
-    def sky_flux(self, sqm: float) -> Optional[float]:
-        """
-        Calculates the sky background flux in electrons per second per pixel.
-        Assumes an L-band (clear) filter equivalent and a zero point where a 21.83 mag/arcsec^2
-        sky produces a specific reference flux.
-
-        :param sqm: Sky Quality Meter reading in mag/arcsec^2.
-        :return: Sky flux in e-/s/pixel.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        scale = self.pixel_scale()
-        if scale is None or not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-        if self.output.quantum_efficiency is None:
-            return None
-
-        area_cm2 = self.telescope.aperture_area().to("cm**2").magnitude
-        qe = self.output.quantum_efficiency
-        transmission = 1.0
-        for f in self.filters:
-            transmission *= f.transmission
-
-        flux = optics_utils.calculate_sky_flux(
-            sqm, area_cm2, qe, scale.magnitude, transmission
-        )
-        return float(flux)
-
-    def object_flux(
-        self,
-        magnitude: float,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional[float]:
-        """
-        Calculates the total integrated flux from an object in electrons per second.
-        The calculation is based on the object's magnitude and the setup's aperture and QE.
-
-        :param magnitude: Apparent magnitude of the object.
-        :param altitude: Optional altitude in degrees to account for atmospheric extinction.
-        :param extinction_k: Atmospheric extinction coefficient (default 0.2).
-        :return: Total integrated object flux in e-/s.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if (
-            not isinstance(self.output, (Camera, SmartTelescope))
-            or self.output.quantum_efficiency is None
-        ):
-            return None
-
-        # Apply atmospheric extinction if altitude is provided
-        effective_magnitude = magnitude
-        if altitude is not None:
-            effective_magnitude = self.atmospheric_extinction(
-                magnitude, altitude, extinction_k
-            )
-
-        area_cm2 = self.telescope.aperture_area().to("cm**2").magnitude
-        qe = self.output.quantum_efficiency
-        transmission = 1.0
-        for f in self.filters:
-            transmission *= f.transmission
-
-        flux = optics_utils.calculate_object_flux(
-            effective_magnitude, area_cm2, qe, transmission
-        )
-        return float(flux)
-
-    def snr(
-        self,
-        magnitude: float,
-        sqm: float,
-        exposure_time: float,
-        n_subs: int = 1,
-        n_pix: int = 4,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-        in_db: bool = False,
-    ) -> Optional[float]:
-        """
-        Calculates the Signal-to-Noise Ratio (SNR) for an object in the given optical path.
-        The model accounts for object shot noise, sky background shot noise, and sensor read noise.
-        It assumes aperture photometry where the object signal is integrated over `n_pix` pixels.
-
-        :param magnitude: Apparent magnitude of the object.
-        :param sqm: Sky Quality Meter reading in mag/arcsec^2.
-        :param exposure_time: Exposure time per sub-exposure in seconds.
-        :param n_subs: Number of sub-exposures stacked.
-        :param n_pix: Number of pixels over which the object signal is integrated (default 4).
-        :param altitude: Optional altitude in degrees for atmospheric extinction.
-        :param extinction_k: Atmospheric extinction coefficient (default 0.2).
-        :param in_db: If True, returns SNR in decibels (20 * log10).
-        :return: SNR (dimensionless linear ratio or dB).
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        obj_flux = self.object_flux(
-            magnitude, altitude=altitude, extinction_k=extinction_k
-        )
-        sky_flux_val = self.sky_flux(sqm)
-
-        if (
-            obj_flux is None
-            or sky_flux_val is None
-            or not isinstance(self.output, (Camera, SmartTelescope))
-            or self.output.read_noise is None
-        ):
-            return None
-
-        snr_val = optics_utils.calculate_snr(
-            obj_flux,
-            sky_flux_val,
-            self.output.read_noise,
-            exposure_time,
-            n_subs,
-            n_pix,
-            in_db,
-        )
-        return float(snr_val)
-
-    def required_subs_for_snr(
-        self,
-        target_snr: float,
-        magnitude: float,
-        sqm: float,
-        exposure_time: float,
-        n_pix: int = 4,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional[float]:
-        """
-        Calculates the number of sub-exposures needed to reach a target SNR.
-        Formula derived from SNR equation: N = SNR^2 * (S + B + R) / S^2
-        Where S is signal, B is sky noise squared, R is read noise squared (all per sub).
-        """
-        obj_flux = self.object_flux(
-            magnitude, altitude=altitude, extinction_k=extinction_k
-        )
-        sky_flux_val = self.sky_flux(sqm)
-
-        if (
-            obj_flux is None
-            or sky_flux_val is None
-            or not hasattr(self.output, "read_noise")
-            or self.output.read_noise is None
-        ):
-            return None
-
-        # Signal and noise per single sub
-        s = obj_flux * exposure_time
-        b = sky_flux_val * exposure_time * n_pix
-        r = (self.output.read_noise**2) * n_pix
-
-        if s <= 0:
-            return numpy.inf
-
-        n_required = (target_snr**2) * (s + b + r) / (s**2)
-        return float(numpy.ceil(n_required))
-
-    def required_integration_time(
-        self,
-        target_snr: float,
-        magnitude: float,
-        sqm: float,
-        exposure_time: float,
-        n_pix: int = 4,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the total integration time needed to reach a target SNR.
-        """
-        n_subs = self.required_subs_for_snr(
-            target_snr,
-            magnitude,
-            sqm,
-            exposure_time,
-            n_pix=n_pix,
-            altitude=altitude,
-            extinction_k=extinction_k,
-        )
-        if n_subs is None:
-            return None
-        return n_subs * exposure_time * get_unit_registry().second
-
-    def optimum_sub_exposure(
-        self, sqm: float, swamp_factor: float = 10.0
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the exposure time at which the sky noise (shot noise from the sky background)
-        becomes dominant over the sensor's read noise by a given 'swamp factor'.
-
-        A common rule of thumb is a swamp factor of 10.0, meaning sky noise power is 10x
-        the read noise power (sky noise is ~3.16x read noise), resulting in a
-        signal-to-noise ratio loss of only ~5% compared to an infinite exposure.
-
-        Formula: t = (swamp_factor * read_noise^2) / sky_flux
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        flux = self.sky_flux(sqm)
-        if (
-            flux is None
-            or not isinstance(self.output, (Camera, SmartTelescope))
-            or self.output.read_noise is None
-        ):
-            return None
-        # Time where SkyNoise^2 = swamp_factor * ReadNoise^2
-        # SkyNoise^2 = Flux * Time
-        # Time = swamp_factor * ReadNoise^2 / Flux
-        time = (swamp_factor * self.output.read_noise**2) / flux
-        return time * get_unit_registry().second
-
-    def camera_limiting_magnitude(
-        self,
-        sqm: float,
-        total_integration_time: float,
-        sub_exposure_time: float,
-        target_snr: float = 5.0,
-        n_pix: int = 4,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional[float]:
-        """
-        Calculates the limiting magnitude for a camera based on reaching a target SNR.
-        Uses binary search to find the magnitude where SNR equals target_snr.
-        Search range: 0.0 to 30.0 magnitude. Convergence: 0.01 magnitude.
-        """
-        low = 0.0
-        high = 30.0
-        n_subs = int(numpy.ceil(total_integration_time / sub_exposure_time))
-
-        for _ in range(12):  # 2^12 = 4096, plenty for 0.01 prec in 30 range
-            mid = (low + high) / 2
-            current_snr = self.snr(
-                mid,
-                sqm,
-                sub_exposure_time,
-                n_subs=n_subs,
-                n_pix=n_pix,
-                altitude=altitude,
-                extinction_k=extinction_k,
-            )
-            if current_snr is None:
-                return None
-            if current_snr > target_snr:
-                low = mid
-            else:
-                high = mid
-
-        return round(float(low), 2)
-
-    def limiting_magnitude(self, sqm: float, integration_time: float) -> float:
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        # integration_time in seconds
-        if (
-            not isinstance(self.output, (Camera, SmartTelescope))
-            or self.output.quantum_efficiency is None
-        ):
-            # Fallback to telescope limiting magnitude if no camera data
-            return self.telescope.limiting_magnitude()
-
-        aperture_cm = self.telescope.aperture.to("cm").magnitude
-        # Base limiting magnitude for 1 second at SQM 21
-        base = 7.7 + 5 * numpy.log10(aperture_cm)
-        # Add integration time factor: 1.25 * log10(T)
-        time_factor = 1.25 * numpy.log10(integration_time)
-        # SQM factor: (SQM - 21)
-        sqm_factor = sqm - 21.0
-
-        return base + time_factor + sqm_factor
-
-    def npf_rule(
-        self, declination: float = 0, k: float = 1.0, simplified: bool = False
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the maximum exposure time to avoid star trailing using the NPF rule.
-        The NPF rule is more accurate than the 'Rule of 500' for modern high-resolution sensors.
-
-        Two versions are available:
-        1. Complete (default): t = k * (16.9 * N + 0.10 * F + 13.7 * P) / (F * cos(dec))
-        2. Simplified: t = k * (35 * N + 30 * P) / (F * cos(dec))
-
-        Where:
-        - N: f-number (focal ratio)
-        - P: pixel pitch in micrometers (µm)
-        - F: focal length in millimeters (mm)
-        - dec: declination of the target in degrees
-        - k: tolerance factor (1.0 for pinpoint stars, up to 3.0 for slightly elongated)
-
-        Source: Frédéric Michaud, Société Astronomique du Havre (SAH)
-        https://sahavre.fr/wp/regle-npf-rule/
-        """
-        if not hasattr(self.output, "pixel_size") or not hasattr(
-            self.telescope, "focal_ratio"
-        ):
-            return None
-
-        # F-number (N)
-        n = (self.telescope.focal_ratio() * self.effective_barlow()).magnitude
-        # Pixel pitch in microns (P)
-        p_size_q = self.output.pixel_size()
-        if p_size_q is None:
-            return None
-        p = p_size_q.to("micrometer").magnitude
-        # Focal length in mm (F)
-        f = (self.telescope.focal_length * self.effective_barlow()).to("mm").magnitude
-
-        res = optics_utils.calculate_npf_rule(n, p, f, declination, k, simplified)
-        return float(res) * get_unit_registry().second
-
-    def estimated_star_trailing(
-        self, exposure_time: float, declination: float = 0
-    ) -> Optional[float]:
-        """
-        Estimates the amount of star trailing in pixels for a given exposure time and declination
-        assuming a stationary (non-tracking) mount.
-
-        Formula: trailing_pixels = (sidereal_rate * exposure_time * cos(declination)) / pixel_scale
-        Where sidereal_rate is ~15.041 arcseconds per second.
-        """
-        scale_q = self.pixel_scale()
-        if scale_q is None:
-            return None
-
-        scale = scale_q.magnitude  # arcsec/pixel
-        if scale == 0:
-            return 0.0
-
-        # Sidereal rotation rate in arcseconds per second
-        sidereal_rate = 15.041067
-
-        cos_dec = math.cos(math.radians(declination))
-        trailing_arcsec = sidereal_rate * exposure_time * cos_dec
-
-        return float(trailing_arcsec / scale)
-
-    def field_rotation_rate(
-        self, latitude_deg: float, azimuth_deg: float, altitude_deg: float
-    ) -> "Quantity":
-        """
-        Calculates the field rotation rate for an Alt-Az mount in arcseconds per second.
-        Formula: omega_rot = omega_earth * cos(lat) * cos(az) / cos(alt)
-        Where omega_earth is the sidereal rotation rate (15.041067 "/s).
-        Source: "Field Rotation" - Bill Keicher
-        """
-        rate = optics_utils.calculate_field_rotation_rate(
-            latitude_deg, azimuth_deg, altitude_deg
-        )
-        return float(rate) * (
-            get_unit_registry().arcsecond / get_unit_registry().second
-        )
-
-    def max_exposure_alt_az(
-        self,
-        latitude_deg: float,
-        azimuth_deg: float,
-        altitude_deg: float,
-        tolerance_pixels: float = 1.0,
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the maximum exposure time for an Alt-Az mount to avoid field rotation trailing.
-        The calculation is based on the movement of the furthest pixel from the sensor center (the corners).
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-
-        # Field rotation rate in arcsec/s
-        rot_rate_q = self.field_rotation_rate(latitude_deg, azimuth_deg, altitude_deg)
-        if rot_rate_q is None:
-            return None
-        rot_rate = rot_rate_q.to("arcsecond/second").magnitude
-
-        if rot_rate < 1e-10:
-            return 3600 * get_unit_registry().second
-
-        # Pixel scale in arcsec/pixel
-        p_scale = self.pixel_scale()
-        if p_scale is None:
-            return None
-
-        # Distance from center to corner in pixels
-        # r = sqrt((width/2)^2 + (height/2)^2)
-        r = 0.5 * numpy.sqrt(self.output.width**2 + self.output.height**2)
-
-        t = (tolerance_pixels * astronomy.RAD_TO_ARCSEC) / (r * rot_rate)
-
-        return t * get_unit_registry().second
-
-    def dawes_limit(self) -> Optional["Quantity"]:
-        """
-        Calculates the Dawes' limit (resolving power) of the telescope in arcseconds.
-        Based on the telescope aperture.
-        """
-        if "dawes_limit" not in self._cache:
-            if hasattr(self.telescope, "dawes_limit"):
-                self._cache["dawes_limit"] = self.telescope.dawes_limit()
-            else:
-                self._cache["dawes_limit"] = None
-        return self._cache["dawes_limit"]
-
-    def rayleigh_limit(self, wavelength_nm: float | int = 550) -> Optional["Quantity"]:
-        """
-        Calculates the Rayleigh limit (resolving power) of the telescope in arcseconds.
-        Based on the telescope aperture and the provided wavelength (default 550nm).
-        """
-        cache_key = f"rayleigh_limit_{wavelength_nm}"
-        if cache_key not in self._cache:
-            if hasattr(self.telescope, "rayleigh_limit"):
-                self._cache[cache_key] = self.telescope.rayleigh_limit(
-                    wavelength_nm=wavelength_nm
-                )
-            else:
-                self._cache[cache_key] = None
-        return self._cache[cache_key]
-
-    def airy_disk_diameter(self, wavelength_nm: float = 550) -> Optional["Quantity"]:
-        """
-        Calculates the physical diameter of the Airy disk (first dark ring) in micrometers.
-        Formula: D = 2.44 * lambda * f/D
-        Where lambda is the wavelength and f/D is the effective focal ratio.
-        This represents the diffraction-limited spot size on the focal plane.
-        """
-        if not hasattr(self.telescope, "focal_ratio"):
-            return None
-        # Effective focal ratio
-        fr = (self.telescope.focal_ratio() * self.effective_barlow()).magnitude
-        diameter_um = optics_utils.calculate_airy_disk_diameter(fr, wavelength_nm)
-        return float(diameter_um) * get_unit_registry().micrometer
-
-    def ideal_planetary_focal_ratio(self, k: float = 5.0) -> Optional[float]:
-        """
-        Calculates the ideal focal ratio for planetary imaging based on the pixel size.
-        Rule of thumb: Ideal focal ratio is between 3x and 5x the pixel size in microns.
-        Formula: Ideal Focal Ratio = k * Pixel Size (µm)
-        Where k is usually 5.0 for average to good seeing.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-
-        # Pixel size in microns
-        p_size_q = self.output.pixel_size()
-        if p_size_q is None:
-            return None
-        p_size = p_size_q.to("micrometer").magnitude
-        return k * p_size
-
-    def nyquist_focal_ratio(
-        self, wavelength_nm: float = 550, sampling_factor: float = 3.0
-    ) -> Optional[float]:
-        """
-        Calculates the ideal focal ratio for a given wavelength and sampling factor based on the Nyquist criterion.
-        Formula: f/D = (p * s) / (1.22 * lambda)
-        Where p is pixel size (µm), s is sampling factor, and lambda is wavelength (µm).
-        A sampling factor of 2.0-3.0 is typically used for planetary imaging.
-        Source: Nyquist-Shannon sampling theorem applied to diffraction-limited optics.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-
-        p_size_q = self.output.pixel_size()
-        if p_size_q is None:
-            return None
-
-        p_um = p_size_q.to("micrometer").magnitude
-        lambda_um = wavelength_nm / 1000.0
-
-        return (p_um * sampling_factor) / (1.22 * lambda_um)
-
-    def planetary_size_in_pixels(
-        self, planet_name: str, time: Any, which: str = "equatorial"
-    ) -> Optional[Union[float, numpy.ndarray]]:
-        """
-        Calculates the projected size of a planet on the sensor in pixels.
-        Uses the planet's angular diameter and the setup's pixel scale.
-
-        :param planet_name: Name of the planet (e.g., 'Jupiter').
-        :param time: Skyfield Time object.
-        :param which: 'equatorial' (default), 'polar', or 'apparent_polar'.
-                      'apparent_polar' accounts for the planet's tilt towards Earth.
-        """
-        from .utils import planetary
-
-        angular_diameter = planetary.get_planet_angular_diameter(
-            planet_name, time, which=which
-        )
-        p_scale = self.pixel_scale()
-
-        if p_scale is None or p_scale.magnitude == 0:
-            return None
-
-        res = angular_diameter / p_scale.magnitude
-        if numpy.isscalar(res):
-            return float(cast(Any, res))
-        return cast(numpy.ndarray, res)
-
-    def saturn_ring_size_in_pixels(
-        self, time: Any
-    ) -> Optional[tuple[Union[float, numpy.ndarray], Union[float, numpy.ndarray]]]:
-        """
-        Calculates the projected size of Saturn's rings on the sensor in pixels.
-        Returns a tuple (major_axis_pixels, minor_axis_pixels).
-        """
-        from .utils import planetary
-
-        details = planetary.get_saturn_ring_details(time)
-        p_scale = self.pixel_scale()
-
-        if p_scale is None or p_scale.magnitude == 0:
-            return None
-
-        major = details["major_axis_arcsec"] / p_scale.magnitude
-        minor = details["minor_axis_arcsec"] / p_scale.magnitude
-
-        def _maybe_cast(val):
-            if numpy.isscalar(val):
-                return float(cast(Any, val))
-            return cast(numpy.ndarray, val)
-
-        return _maybe_cast(major), _maybe_cast(minor)
-
-    def rule_of_500(self) -> "Quantity":
-        """
-        Calculates the maximum exposure time to avoid star trailing using the classic Rule of 500.
-        Formula: t = 500 / (F_actual * crop_factor)
-        """
-        # Actual focal length
-        f_actual = (
-            (self.telescope.focal_length * self.effective_barlow()).to("mm").magnitude
-        )
-
-        if f_actual == 0:
-            return 0 * get_unit_registry().second
-
-        if hasattr(self.output, "sensor_width") and hasattr(
-            self.output, "sensor_height"
-        ):
-            # crop_factor = 43.27 / diagonal
-            # diagonal of 35mm sensor (36x24) is ~43.27mm
-            diagonal = numpy.sqrt(
-                self.output.sensor_width.to("mm").magnitude ** 2
-                + self.output.sensor_height.to("mm").magnitude ** 2
-            )
-            if diagonal == 0:
-                return (500 / f_actual) * get_unit_registry().second
-            crop_factor = 43.27 / diagonal
-            t = 500 / (f_actual * crop_factor)
-        else:
-            # Fallback for non-camera or visual setup (though less relevant)
-            t = 500 / f_actual
-
-        return t * get_unit_registry().second
-
-    def dynamic_range(self) -> Optional[float]:
-        """
-        Calculates the dynamic range of the camera or smart telescope in the optical path.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if isinstance(self.output, (Camera, SmartTelescope)):
-            return self.output.dynamic_range()
-        return None
-
-    def psf_peak_fraction(self, seeing: float) -> Optional[float]:
-        """
-        Calculates the fraction of light from a point source that falls into the central pixel,
-        assuming a Gaussian Point Spread Function (PSF).
-        Formula: f = erf(pixel_scale * sqrt(ln2) / seeing)^2
-        Where 'seeing' is the Full Width at Half Maximum (FWHM) in arcseconds.
-        """
-        scale = self.pixel_scale()
-        if scale is None or seeing <= 0:
-            return None
-
-        # f = erf(L * sqrt(ln2) / FWHM)^2
-        # where L is pixel scale and FWHM is seeing.
-        arg = (scale.magnitude * math.sqrt(math.log(2))) / seeing
-        fraction = math.erf(arg) ** 2
-        return float(fraction)
-
-    def saturation_time(
-        self,
-        magnitude: float,
-        seeing: float,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the maximum exposure time before the central pixel of a point source saturates.
-        Based on the full-well capacity of the sensor and the Gaussian PSF model.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-
-        if self.output.full_well is None:
-            return None
-
-        obj_flux = self.object_flux(
-            magnitude, altitude=altitude, extinction_k=extinction_k
-        )
-        f = self.psf_peak_fraction(seeing)
-
-        if obj_flux is None or f is None or obj_flux <= 0 or f <= 0:
-            return None
-
-        # S_peak = flux * t * f
-        # t = full_well / (flux * f)
-        t = self.output.full_well / (obj_flux * f)
-        return t * get_unit_registry().second
-
-    def saturation_magnitude(
-        self,
-        exposure_time: float,
-        seeing: float,
-        altitude: Optional[float] = None,
-        extinction_k: float = 0.2,
-    ) -> Optional[float]:
-        """
-        Calculates the magnitude of a point source that would just saturate the sensor
-        at the given exposure time and seeing conditions.
-        """
-        from .opticalequipment.camera import Camera
-        from .opticalequipment.smart_telescope import SmartTelescope
-
-        if not isinstance(self.output, (Camera, SmartTelescope)):
-            return None
-
-        if self.output.full_well is None:
-            return None
-
-        f = self.psf_peak_fraction(seeing)
-        if f is None or f <= 0 or exposure_time <= 0:
-            return None
-
-        # target_flux = full_well / (t * f)
-        target_flux = self.output.full_well / (exposure_time * f)
-
-        # flux = 0.005 * 10^(0.4 * (21.83 - m_eff)) * Area * QE * Transmission
-        # We can find m_eff using binary search or analytical inversion.
-        # Since we have object_flux(m), we know object_flux(0)
-        flux_at_zero = self.object_flux(0.0, altitude=None, extinction_k=0.0)
-        if flux_at_zero is None:
-            return None
-
-        # target_flux = flux_at_zero * 10^(-0.4 * m_eff)
-        # 10^(-0.4 * m_eff) = target_flux / flux_at_zero
-        # -0.4 * m_eff = log10(target_flux / flux_at_zero)
-        # m_eff = -2.5 * log10(target_flux / flux_at_zero)
-        m_eff = -2.5 * math.log10(target_flux / flux_at_zero)
-
-        if altitude is not None:
-            # m_eff = m + k * X
-            # m = m_eff - k * X
-            airmass_val = self.airmass(altitude)
-            return float(m_eff - extinction_k * airmass_val)
-
-        return float(m_eff)
-
-    def planetary_phase_angle(
-        self, planet_name: str, time: Any
-    ) -> Union[float, numpy.ndarray]:
-        """
-        Calculates the phase angle (Sun-Object-Earth) for a planet or the Moon.
-        """
-        from .utils import planetary
-
-        return planetary.get_planet_phase_angle(planet_name, time)
-
-    def moon_libration(self, time: Any) -> tuple[float, float]:
-        """
-        Calculates the Moon's libration in longitude and latitude in degrees.
-        """
-        from .utils import planetary
-
-        lon, lat = planetary.get_moon_libration(time)
-        if numpy.isscalar(lon):
-            return float(cast(Any, lon)), float(cast(Any, lat))
-        return cast(Any, (lon, lat))
-
-    def moon_position_angle_bright_limb(self, time: Any) -> float:
-        """
-        Calculates the position angle of the Moon's bright limb in degrees.
-        """
-        from .utils import planetary
-
-        return planetary.get_moon_position_angle_bright_limb(time)
-
-    def planetary_magnitude(
-        self, planet_name: str, time: Any
-    ) -> Union[float, numpy.ndarray]:
-        """
-        Calculates the apparent magnitude of a planet, the Moon, or the Sun.
-        """
-        from .utils import planetary
-
-        return planetary.get_planet_magnitude(planet_name, time)
-
-    def planetary_phase(
-        self, planet_name: str, time: Any
-    ) -> Union[float, numpy.ndarray]:
-        """
-        Calculates the illuminated fraction of a planet or the Moon as a percentage (0-100).
-        """
-        from .utils import planetary
-
-        return planetary.get_planet_phase(planet_name, time)
-
-    def planetary_surface_brightness(
-        self, planet_name: str, time: Any
-    ) -> Union[float, numpy.ndarray]:
-        """
-        Calculates the average surface brightness of a planet, the Moon, or the Sun
-        in mag/arcsec².
-        """
-        from .utils import planetary
-
-        return planetary.get_planet_surface_brightness(planet_name, time)
-
-    def moon_colongitude(self, time: Any) -> Union[float, numpy.ndarray]:
-        """
-        Calculates the Moon's selenographic colongitude in degrees.
-        """
-        from .utils import planetary
-
-        return planetary.get_moon_colongitude(time)
-
-    def jupiter_cml(self, time: Any, system: int = 2) -> Union[float, numpy.ndarray]:
-        """
-        Calculates Jupiter's Central Meridian Longitude (CML) for the specified system (1 or 2).
-        """
-        from .utils import planetary
-
-        return planetary.get_jupiter_cml(time, system)
-
-    def mars_cml(self, time: Any) -> Union[float, numpy.ndarray]:
-        """
-        Calculates Mars' Central Meridian Longitude (CML).
-        """
-        from .utils import planetary
-
-        return planetary.get_mars_cml(time)
-
-    def saturn_cml(self, time: Any, system: int = 3) -> Union[float, numpy.ndarray]:
-        """
-        Calculates Saturn's Central Meridian Longitude (CML) for the specified system (1, 2 or 3).
-        """
-        from .utils import planetary
-
-        return planetary.get_saturn_cml(time, system)
-
-    def sun_physical_details(self, time: Any) -> dict:
-        """
-        Calculates the physical details of the Sun (P, B0, L0).
-        """
-        from .utils import planetary
-
-        return planetary.get_sun_physical_details(time)
-
-    def max_planetary_rotation_duration(
-        self, planet_name: str, time: Any, tolerance_pixels: float = 1.0
-    ) -> Optional["Quantity"]:
-        """
-        Calculates the maximum recording duration (in seconds) before planetary rotation
-        causes a blur exceeding the given pixel tolerance.
-        Useful for planetary imagers (lucky imaging) to determine when to start a new capture
-        or use tools like WinJUPOS for de-rotation.
-        """
-        from .utils import planetary
-
-        scale_q = self.pixel_scale()
-        if scale_q is None:
-            return None
-
-        scale = scale_q.to("arcsecond").magnitude
-
-        # Equatorial angular radius (half of diameter)
-        r_eq = (
-            planetary.get_planet_angular_diameter(planet_name, time, which="equatorial")
-            / 2.0
-        )
-
-        # Sidereal rotation period in seconds
-        period = planetary.get_planet_rotation_period(planet_name)
-
-        # Sub-observer latitude (tilt)
-        try:
-            de_deg = planetary.get_sub_observer_latitude(planet_name, time)
-            cos_de = math.cos(math.radians(de_deg))
-        except ValueError:
-            # Fallback for planets without pole models (e.g., Mercury, Venus)
-            cos_de = 1.0
-
-        if r_eq <= 0 or period <= 0:
-            return None
-
-        # Angular velocity of a point at the center of the disk as seen from Earth (arcsec/s)
-        # v_arcsec = (2 * pi * r_eq_arcsec * cos(De)) / T
-        omega = (2.0 * math.pi * r_eq * cos_de) / period
-
-        if omega <= 1e-12:
-            return (
-                3600 * get_unit_registry().second
-            )  # Cap at 1 hour for very slow rotators
-
-        t_max = (tolerance_pixels * scale) / omega
-
-        return t_max * get_unit_registry().second
