+++ apts/optics/__init__.py
@@ -1,0 +1,4 @@
+from .base import OpticalPath
+from .utils import OpticsUtils
+
+__all__ = ["OpticalPath", "OpticsUtils"]
